requirements.txt
absl-py==0.9.0
adal==1.2.4
aiohttp==3.6.2
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
applicationinsights==0.11.9
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.1
astroid==2.4.0
astropy==4.0.1.post1
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
azure-common==1.1.25
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-network==10.2.0
azure-mgmt-resource==10.0.0
azure-mgmt-storage==11.1.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azureml-core==1.6.0
azureml-telemetry==1.6.0
azureml-train-core==1.6.0
azureml-train-restclients-hyperdrive==1.6.0
Babel==2.8.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.0
bert-for-tf2==0.13.5
bitarray==1.2.1
bitstring==3.1.7
bkcharts==0.2
bleach==3.1.4
blobxfer==1.9.4.post1
bokeh==2.0.2
boto==2.49.0
boto3==1.13.24
botocore==1.16.24
Bottleneck==1.3.2
bz2file==0.98
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-completion==0.5.2
click-option-group==0.4.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.3
colorlog==4.1.0
configparser==5.0.0
contextlib2==0.6.0.post1
contextvars==2.4
coverage==5.1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.16.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.1.1
distributed==2.16.0
docker==3.7.3
docker-pycreds==0.4.0
docutils==0.15.2
dopamine-rl==3.0.1
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.5.0
flake8-mypy==17.8.0
flashtool==0.0.7
Flask==1.1.2
fsspec==0.7.1
future==0.18.2
gast==0.2.2
gevent==1.4.0
gin-config==0.3.0
gitdb==4.0.5
GitPython==3.1.2
glob2==0.7
gmpy2==2.0.8
google-api-core==1.19.1
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
gql==0.2.0
graphql-core==1.1
greenlet==0.4.15
grpcio==1.29.0
gunicorn==20.0.4
gym==0.14.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.18.1
humanize==2.4.0
hypothesis==5.11.0
idna==2.6
idna-ssl==1.1.0
imageio==2.8.0
imagesize==1.2.0
immutables==0.11
importlib-metadata==1.6.1
importlib-resources==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.4
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.1.1
kfac==0.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.32.1
locket==0.2.0
lxml==4.5.0
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.5.0
matplotlib==3.1.3
mccabe==0.6.1
mesh-tensorflow==0.1.13
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.0
msrest==0.6.17
msrestazure==0.6.4
multidict==4.7.6
multipledispatch==0.6.0
mypy==0.780
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.6
ndg-httpsclient==0.5.1
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
ntlm-auth==1.4.0
numba==0.49.1
numexpr==2.7.1
numpy==1.18.5
numpydoc==0.9.2
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
params-flow==0.7.4
parso==0.5.2
partd==1.1.0
passpy==1.0.2
path==13.1.0
pathlib2==2.3.5
pathspec==0.8.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
phillytools==6.0.2
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
portalocker==1.7.0
progress==1.5
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
py-params==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pyflakes==1.6.0
pyglet==1.3.2
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pypng==0.0.20
pyrsistent==0.16.0
PySocks==1.7.1
pytest==3.2.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-cov==2.5.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-gnupg==0.4.6
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python-Levenshtein==0.12.0
-e git+https://github.com/kmkurn/pytorch-crf.git@4cd79bc8af55fb0f34a2a39b2e38f0e71c208fd4#egg=pytorch_crf
pytorch-lightning==0.7.3
pytorch-pretrained-bert==0.4.0
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.5.14
requests==2.22.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
rope==0.17.0
rouge-score==0.0.3
rsa==4.0
Rtree==0.9.4
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sacrebleu==1.4.9
sacremoses==0.0.43
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
SecretStorage==3.1.2
Send2Trash==1.5.0
sentence-transformers==0.2.6.1
sentencepiece==0.1.91
sentry-sdk==0.15.1
seqeval==0.0.12
shellingham==1.3.2
shortuuid==1.0.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
statsmodels==0.11.0
subprocess32==3.5.4
sympy==1.5.1
tables==3.6.1
tabulate==0.8.7
tblib==1.6.0
tensor2tensor==1.15.6
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.15.0
tensorflow-addons==0.10.0
tensorflow-datasets==3.1.0
tensorflow-estimator==1.15.1
tensorflow-gan==2.0.0
tensorflow-gpu==2.0.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.0
tensorflow-probability==0.7.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-slim==1.1.0
tokenizers==0.7.0
toml==0.10.0
toolz==0.10.0
torch==1.5.0
TorchCRF==1.0.6
torchvision==0.6.0a0+82fd1c8
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
transformers==2.10.0
typed-ast==1.4.1
typeguard==2.9.1
typing==3.7.4.1
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
wandb==0.9.1
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xdg==4.0.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
yapf==0.25.0
yarl==1.4.2
zict==2.0.0
zipp==3.1.0
